EHR/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.5
web3==6.0.0
PyJWT==2.6.0
cryptography==39.0.1
requests==2.28.2
scikit-learn==1.2.2
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
pytz==2022.7.1
psycopg2-binary==2.9.5
# Replace the Hyperledger Fabric SDK with a more compatible version
# fabric-sdk-py==0.9.0
# hyperledger-fabric==0.9.0
gunicorn==20.1.0
pytest==7.3.1 